Build the first cn C file
