+++ linker_scripts/cn/symbol_addrs_dma_table.txt
@@ -1,0 +1,1 @@
+gDmaTable = 0x8002A7B0; // defined:true size:0x10
