+++ src/main_segment/053610.c
@@ -4,10 +4,10 @@
 #include "unknown_structs.h"
 #include "unk.h"
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_ASM("asm/us/nonmatchings/main_segment/053610", func_8007D260);
 #endif
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_ASM("asm/us/nonmatchings/main_segment/053610", func_8007D348);
 #endif
