+++ src/main_segment/dm_virus_init.c
@@ -30,7 +30,7 @@ void init_map_all(GameMapGrid *mapGrid) {
     }
 }
 
-#if VERSION == VERSION_US
+#if VERSION_US
 #ifdef NON_EQUIVALENT
 // not sure if equivalent
 void clear_map(GameMapGrid *mapGrid, s32 column, s32 row) {
@@ -69,15 +69,15 @@ bool get_map_info(GameMapGrid *mapGrid, s32 column, s32 rowPlusOne) {
     return cells[index].unk_4[0] != 0;
 }
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_ASM("asm/us/nonmatchings/main_segment/dm_virus_init", func_8005F09C);
 #endif
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_ASM("asm/us/nonmatchings/main_segment/dm_virus_init", func_8005F0E4);
 #endif
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_RODATA("asm/us/nonmatchings/main_segment/dm_virus_init", D_800B1BA0);
 #endif
 
@@ -176,7 +176,7 @@ u32 func_8005F310(struct_virus_map_data *arg0, s32 arg1, s32 arg2) {
     return (~value) >> 31;
 }
 
-#if VERSION == VERSION_US
+#if VERSION_US
 #ifdef NON_MATCHING
 bool dm_check_color(struct_virus_map_data *arg0, u16 arg1, s32 arg2, u8 arg3) {
     struct {
@@ -329,14 +329,14 @@ s32 func_8005F6F0(s32 arg0) {
     return D_800B1BC8[arg0];
 }
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_RODATA("asm/us/nonmatchings/main_segment/dm_virus_init", D_800B1BC8);
 #endif
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_RODATA("asm/us/nonmatchings/main_segment/dm_virus_init", D_800B1BCC);
 #endif
 
-#if VERSION == VERSION_US
+#if VERSION_US
 #if 0
 extern u8 D_800B1BCC[];
 extern u8 D_800B1BA0[];
@@ -534,15 +534,15 @@ void func_8005FC6C(struct_8005FC6C_arg0 *arg0, s32 arg1, s32 arg2, s32 arg3) {
     arg0->unk_008 = arg3;
 }
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_ASM("asm/us/nonmatchings/main_segment/dm_virus_init", _makeFlash_checkOrdre);
 #endif
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_ASM("asm/us/nonmatchings/main_segment/dm_virus_init", func_8005FE68);
 #endif
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_RODATA("asm/us/nonmatchings/main_segment/dm_virus_init", D_800B1BD0);
 #endif
 
@@ -554,6 +554,6 @@ struct _m2c_stack_make_flash_virus_pos {
     /* 0x310 */ char pad_310[0x30];
 }; /* size = 0x340 */
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_ASM("asm/us/nonmatchings/main_segment/dm_virus_init", make_flash_virus_pos);
 #endif
