+++ src/main_segment/vr_init.c
@@ -17,6 +17,6 @@ void sgenrand(u32 arg0) {
     }
 }
 
-#if VERSION == VERSION_US
+#if VERSION_US
 INCLUDE_ASM("asm/us/nonmatchings/main_segment/vr_init", genrand);
 #endif
