+++ tools/.gitignore
@@ -1,1 +1,2 @@
 gcc_kmc/
+gcc_egcs/
